requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# Web Search & Scraping
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
httpx>=0.25.0
serpapi>=0.1.5

# AI & ML
openai>=1.3.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
pyyaml>=6.0

# Sentiment Analysis
vaderSentiment>=3.3.2

# Visualization
plotly>=5.17.0

# Task Queue (Optional - for advanced features)
celery>=5.3.0
redis>=5.0.0

# Async & Retry
tenacity>=8.2.0

# Testing (Dev)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Database (Optional)
sqlalchemy>=2.0.0

# API Server (Optional - if using Flask API)
flask>=3.0.0
flask-socketio>=5.3.0
eventlet>=0.33.0

# Additional utilities
requests>=2.31.0
tqdm>=4.66.0
